requirements.txt
torch
torchvision
Pillow
fastapi
uvicorn
requests